Some code editing. SafeWrite for functions arguments replaced WinAPI types to C++ standart.

+++ sfall/SafeWrite.cpp
@@ -3,7 +3,7 @@
 namespace sfall
 {
 
-static void __stdcall SafeWriteFunc(BYTE code, DWORD addr, void* func) {
+static void __stdcall SafeWriteFunc(uint8_t code, uint32_t addr, void* func) {
 	DWORD oldProtect, data = (DWORD)func - (addr + 5);
 
 	VirtualProtect((void*)addr, 5, PAGE_EXECUTE_READWRITE, &oldProtect);
@@ -14,7 +14,7 @@ static void __stdcall SafeWriteFunc(BYTE code, DWORD addr, void* func) {
 	CheckConflict(addr, 5);
 }
 
-static __declspec(noinline) void __stdcall SafeWriteFunc(BYTE code, DWORD addr, void* func, DWORD len) {
+static __declspec(noinline) void __stdcall SafeWriteFunc(uint8_t code, uint32_t addr, void* func, size_t len) {
 	DWORD oldProtect,
 		protectLen = len + 5,
 		addrMem = addr + 5,
@@ -24,7 +24,7 @@ static __declspec(noinline) void __stdcall SafeWriteFunc(BYTE code, DWORD addr, 
 	*((BYTE*)addr) = code;
 	*((DWORD*)(addr + 1)) = data;
 
-	for (unsigned int i = 0; i < len; i++) {
+	for (size_t i = 0; i < len; i++) {
 		*((BYTE*)(addrMem + i)) = CodeType::Nop;
 	}
 	VirtualProtect((void*)addr, protectLen, oldProtect, &oldProtect);
@@ -32,7 +32,7 @@ static __declspec(noinline) void __stdcall SafeWriteFunc(BYTE code, DWORD addr, 
 	CheckConflict(addr, protectLen);
 }
 
-void SafeWriteBytes(DWORD addr, BYTE* data, int count) {
+void SafeWriteBytes(uint32_t addr, uint8_t* data, size_t count) {
 	DWORD oldProtect;
 
 	VirtualProtect((void*)addr, count, PAGE_EXECUTE_READWRITE, &oldProtect);
@@ -42,7 +42,7 @@ void SafeWriteBytes(DWORD addr, BYTE* data, int count) {
 	AddrAddToList(addr, count);
 }
 
-void __stdcall SafeWrite8(DWORD addr, BYTE data) {
+void __stdcall SafeWrite8(uint32_t addr, uint8_t data) {
 	DWORD oldProtect;
 
 	VirtualProtect((void*)addr, 1, PAGE_EXECUTE_READWRITE, &oldProtect);
@@ -52,7 +52,7 @@ void __stdcall SafeWrite8(DWORD addr, BYTE data) {
 	AddrAddToList(addr, 1);
 }
 
-void __stdcall SafeWrite16(DWORD addr, WORD data) {
+void __stdcall SafeWrite16(uint32_t addr, uint16_t data) {
 	DWORD oldProtect;
 
 	VirtualProtect((void*)addr, 2, PAGE_EXECUTE_READWRITE, &oldProtect);
@@ -62,7 +62,7 @@ void __stdcall SafeWrite16(DWORD addr, WORD data) {
 	AddrAddToList(addr, 2);
 }
 
-void __stdcall SafeWrite32(DWORD addr, DWORD data) {
+void __stdcall SafeWrite32(uint32_t addr, uint32_t data) {
 	DWORD oldProtect;
 
 	VirtualProtect((void*)addr, 4, PAGE_EXECUTE_READWRITE, &oldProtect);
@@ -72,7 +72,7 @@ void __stdcall SafeWrite32(DWORD addr, DWORD data) {
 	AddrAddToList(addr, 4);
 }
 
-void __stdcall SafeWriteStr(DWORD addr, const char* data) {
+void __stdcall SafeWriteStr(uint32_t addr, const char* data) {
 	DWORD oldProtect;
 	long len = strlen(data) + 1;
 
@@ -83,47 +83,47 @@ void __stdcall SafeWriteStr(DWORD addr, const char* data) {
 	AddrAddToList(addr, len);
 }
 
-void HookCall(DWORD addr, void* func) {
+void HookCall(uint32_t addr, void* func) {
 	SafeWrite32(addr + 1, (DWORD)func - (addr + 5));
 
 	CheckConflict(addr, 1);
 }
 
-void MakeCall(DWORD addr, void* func) {
+void MakeCall(uint32_t addr, void* func) {
 	SafeWriteFunc(CodeType::Call, addr, func);
 }
 
-void MakeCall(DWORD addr, void* func, int len) {
+void MakeCall(uint32_t addr, void* func, size_t len) {
 	SafeWriteFunc(CodeType::Call, addr, func, len);
 }
 
-void MakeJump(DWORD addr, void* func) {
+void MakeJump(uint32_t addr, void* func) {
 	SafeWriteFunc(CodeType::Jump, addr, func);
 }
 
-void MakeJump(DWORD addr, void* func, int len) {
+void MakeJump(uint32_t addr, void* func, size_t len) {
 	SafeWriteFunc(CodeType::Jump, addr, func, len);
 }
 
-void HookCalls(void* func, std::initializer_list<DWORD> addrs) {
+void HookCalls(void* func, std::initializer_list<uint32_t> addrs) {
 	for (auto& addr : addrs) {
 		HookCall(addr, func);
 	}
 }
 
-void MakeCalls(void* func, std::initializer_list<DWORD> addrs) {
+void MakeCalls(void* func, std::initializer_list<uint32_t> addrs) {
 	for (auto& addr : addrs) {
 		MakeCall(addr, func);
 	}
 }
 
-void MakeJumps(void* func, std::initializer_list<DWORD> addrs) {
+void MakeJumps(void* func, std::initializer_list<uint32_t> addrs) {
 	for (auto& addr : addrs) {
 		MakeJump(addr, func);
 	}
 }
 
-void SafeMemSet(DWORD addr, BYTE val, int len) {
+void SafeMemSet(uint32_t addr, uint8_t val, size_t len) {
 	DWORD oldProtect;
 
 	VirtualProtect((void*)addr, len, PAGE_EXECUTE_READWRITE, &oldProtect);
@@ -133,7 +133,7 @@ void SafeMemSet(DWORD addr, BYTE val, int len) {
 	AddrAddToList(addr, len);
 }
 
-void BlockCall(DWORD addr) {
+void BlockCall(uint32_t addr) {
 	DWORD oldProtect;
 
 	VirtualProtect((void*)addr, 4, PAGE_EXECUTE_READWRITE, &oldProtect);
